requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
aiohappyeyeballs==2.4.4
    # via aiohttp
aiohttp==3.11.11
    # via
    #   datasets
    #   fsspec
    #   langchain
    #   langchain-community
aiosignal==1.3.2
    # via aiohttp
altair==5.5.0
    # via streamlit
annotated-types==0.7.0
    # via pydantic
anyio==4.8.0
    # via
    #   httpx
    #   openai
appdirs==1.4.4
    # via ragas
attrs==24.3.0
    # via
    #   aiohttp
    #   jsonschema
    #   referencing
black==24.10.0
    # via ai-agents-cs (pyproject.toml)
blinker==1.9.0
    # via streamlit
boto3==1.35.94
    # via
    #   ai-agents-cs (pyproject.toml)
    #   langchain-aws
botocore==1.35.94
    # via
    #   boto3
    #   s3transfer
cachetools==5.5.0
    # via streamlit
certifi==2024.12.14
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.1
    # via requests
click==8.1.8
    # via
    #   black
    #   streamlit
dataclasses-json==0.6.7
    # via langchain-community
datasets==3.2.0
    # via ragas
dill==0.3.8
    # via
    #   datasets
    #   multiprocess
diskcache==5.6.3
    # via ragas
distro==1.9.0
    # via openai
dnspython==2.7.0
    # via email-validator
email-validator==2.2.0
    # via pydantic
filelock==3.16.1
    # via
    #   datasets
    #   huggingface-hub
frozenlist==1.5.0
    # via
    #   aiohttp
    #   aiosignal
fsspec==2024.9.0
    # via
    #   datasets
    #   huggingface-hub
gitdb==4.0.12
    # via gitpython
gitpython==3.1.44
    # via streamlit
h11==0.14.0
    # via httpcore
httpcore==1.0.7
    # via httpx
httpx==0.28.1
    # via
    #   langgraph-sdk
    #   langsmith
    #   openai
httpx-sse==0.4.0
    # via langchain-community
huggingface-hub==0.27.1
    # via datasets
idna==3.10
    # via
    #   anyio
    #   email-validator
    #   httpx
    #   requests
    #   yarl
jinja2==3.1.5
    # via
    #   altair
    #   pydeck
jiter==0.8.2
    # via openai
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
jsonpatch==1.33
    # via langchain-core
jsonpointer==3.0.0
    # via jsonpatch
jsonschema==4.23.0
    # via altair
jsonschema-specifications==2024.10.1
    # via jsonschema
langchain==0.3.14
    # via
    #   ai-agents-cs (pyproject.toml)
    #   langchain-community
    #   ragas
langchain-aws==0.2.10
    # via ai-agents-cs (pyproject.toml)
langchain-community==0.3.14
    # via ragas
langchain-core==0.3.29
    # via
    #   ai-agents-cs (pyproject.toml)
    #   langchain
    #   langchain-aws
    #   langchain-community
    #   langchain-openai
    #   langchain-text-splitters
    #   langgraph
    #   langgraph-checkpoint
    #   ragas
langchain-openai==0.2.14
    # via
    #   ai-agents-cs (pyproject.toml)
    #   ragas
langchain-text-splitters==0.3.5
    # via langchain
langdetect==1.0.9
    # via ai-agents-cs (pyproject.toml)
langgraph==0.2.61
    # via ai-agents-cs (pyproject.toml)
langgraph-checkpoint==2.0.9
    # via langgraph
langgraph-sdk==0.1.48
    # via langgraph
langsmith==0.2.10
    # via
    #   langchain
    #   langchain-community
    #   langchain-core
markdown-it-py==3.0.0
    # via rich
markupsafe==3.0.2
    # via jinja2
marshmallow==3.24.1
    # via dataclasses-json
mdurl==0.1.2
    # via markdown-it-py
msgpack==1.1.0
    # via langgraph-checkpoint
multidict==6.1.0
    # via
    #   aiohttp
    #   yarl
multiprocess==0.70.16
    # via datasets
mypy-extensions==1.0.0
    # via
    #   black
    #   typing-inspect
narwhals==1.21.1
    # via altair
nest-asyncio==1.6.0
    # via ragas
numpy==2.2.1
    # via
    #   datasets
    #   langchain
    #   langchain-aws
    #   langchain-community
    #   pandas
    #   pydeck
    #   ragas
    #   streamlit
openai==1.59.4
    # via
    #   ai-agents-cs (pyproject.toml)
    #   langchain-openai
    #   ragas
orjson==3.10.13
    # via
    #   langgraph-sdk
    #   langsmith
packaging==24.2
    # via
    #   altair
    #   black
    #   datasets
    #   huggingface-hub
    #   langchain-core
    #   marshmallow
    #   streamlit
pandas==2.2.3
    # via
    #   datasets
    #   streamlit
pathspec==0.12.1
    # via black
pillow==11.1.0
    # via streamlit
platformdirs==4.3.6
    # via black
propcache==0.2.1
    # via
    #   aiohttp
    #   yarl
protobuf==5.29.2
    # via streamlit
psycopg==3.2.3
    # via ai-agents-cs (pyproject.toml)
psycopg-binary==3.2.3
    # via psycopg
pyarrow==18.1.0
    # via
    #   datasets
    #   streamlit
pydantic==2.10.4
    # via
    #   ai-agents-cs (pyproject.toml)
    #   langchain
    #   langchain-aws
    #   langchain-core
    #   langsmith
    #   openai
    #   pydantic-settings
    #   ragas
pydantic-core==2.27.2
    # via pydantic
pydantic-settings==2.7.1
    # via langchain-community
pydeck==0.9.1
    # via streamlit
pygments==2.19.1
    # via rich
pysbd==0.3.4
    # via ragas
python-dateutil==2.9.0.post0
    # via
    #   botocore
    #   pandas
python-dotenv==1.0.1
    # via
    #   ai-agents-cs (pyproject.toml)
    #   pydantic-settings
pytz==2024.2
    # via pandas
pyyaml==6.0.2
    # via
    #   datasets
    #   huggingface-hub
    #   langchain
    #   langchain-community
    #   langchain-core
ragas==0.2.9
    # via ai-agents-cs (pyproject.toml)
referencing==0.35.1
    # via
    #   jsonschema
    #   jsonschema-specifications
regex==2024.11.6
    # via tiktoken
requests==2.32.3
    # via
    #   datasets
    #   huggingface-hub
    #   langchain
    #   langchain-community
    #   langsmith
    #   requests-toolbelt
    #   streamlit
    #   tiktoken
requests-toolbelt==1.0.0
    # via langsmith
rich==13.9.4
    # via streamlit
rpds-py==0.22.3
    # via
    #   jsonschema
    #   referencing
s3transfer==0.10.4
    # via boto3
six==1.17.0
    # via
    #   langdetect
    #   python-dateutil
smmap==5.0.2
    # via gitdb
sniffio==1.3.1
    # via
    #   anyio
    #   openai
sqlalchemy==2.0.36
    # via
    #   langchain
    #   langchain-community
streamlit==1.41.1
    # via ai-agents-cs (pyproject.toml)
tenacity==9.0.0
    # via
    #   langchain
    #   langchain-community
    #   langchain-core
    #   streamlit
tiktoken==0.8.0
    # via
    #   langchain-openai
    #   ragas
toml==0.10.2
    # via streamlit
tornado==6.4.2
    # via streamlit
tqdm==4.67.1
    # via
    #   datasets
    #   huggingface-hub
    #   openai
typing-extensions==4.12.2
    # via
    #   altair
    #   huggingface-hub
    #   langchain-core
    #   openai
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   streamlit
    #   typing-inspect
typing-inspect==0.9.0
    # via dataclasses-json
tzdata==2024.2
    # via pandas
urllib3==2.3.0
    # via
    #   botocore
    #   requests
xxhash==3.5.0
    # via datasets
yarl==1.18.3
    # via aiohttp
